requirements.txt
django
django-filter
djangorestframework
requests
pyyaml
django-cors-headers
python-decouple
Pillow
overpy
requests
